backend/requirements.txt
Django==5.0
djangorestframework==3.15.1

Pillow==10.2.0
numpy==1.26.4

rembg[onnx]==2.0.57
onnxruntime==1.17.1

mysqlclient==2.2.4

celery==5.3.6
redis==5.0.1